+++ examples/v2_get_block_certificates.rs
@@ -1,8 +1,9 @@
-//! Test the `GetFirstBlockEpoch` endpoint.
+//! Test the `GetBlockCertificates` endpoint.
 use anyhow::Context;
 use clap::AppSettings;
-use concordium_rust_sdk::v2;
 use structopt::StructOpt;
+
+use concordium_rust_sdk::v2;
 
 #[derive(StructOpt)]
 struct App {
@@ -12,9 +13,8 @@ struct App {
         default_value = "http://localhost:20000"
     )]
     endpoint: v2::Endpoint,
-
-    #[structopt(long = "epoch", help = "Epoch identifier", default_value = "%5,0")]
-    epoch: v2::EpochIdentifier,
+    #[structopt(long = "block", help = "Block to query", default_value = "lastfinal")]
+    bi:       v2::BlockIdentifier,
 }
 
 #[tokio::main(flavor = "multi_thread")]
@@ -28,7 +28,9 @@ async fn main() -> anyhow::Result<()> {
     let mut client = v2::Client::new(app.endpoint)
         .await
         .context("Cannot connect.")?;
-    let fb = client.get_first_block_epoch(app.epoch).await?;
-    println!("{}", fb);
+
+    let certificates = client.get_block_certificates(&app.bi).await?;
+    println!("{:#?}", certificates);
+
     Ok(())
 }