+++ src/v2/mod.rs
@@ -6,7 +6,7 @@ use crate::{
     endpoints, id,
     id::types::AccountCredentialMessage,
     types::{
-        self, hashes,
+        self, block_certificates, hashes,
         hashes::{BlockHash, TransactionHash, TransactionSignHash},
         smart_contracts::{
             ContractContext, InstanceInfo, InvokeContractResult, ModuleReference, WasmModule,
@@ -2331,6 +2331,28 @@ impl Client {
             .await?;
         let block_hash = extract_metadata(&response)?;
         let response = ChainParameters::try_from(response.into_inner())?;
+        Ok(QueryResponse {
+            block_hash,
+            response,
+        })
+    }
+
+    /// For a non-genesis block, this returns the [`QuorumCertificate`], a
+    /// [`TimeoutCertificate`] (if present) and [`EpochFinalizationEntry`] (if
+    /// present).
+    /// If the block being pointed to is *not* from protocol version 6 or
+    /// above, then [`InvalidArgument`](`tonic::Code::InvalidArgument`)
+    /// is returned.
+    pub async fn get_block_certificates(
+        &mut self,
+        bi: impl IntoBlockIdentifier,
+    ) -> endpoints::QueryResult<QueryResponse<block_certificates::BlockCertificates>> {
+        let response = self
+            .client
+            .get_block_certificates(&bi.into_block_identifier())
+            .await?;
+        let block_hash = extract_metadata(&response)?;
+        let response = block_certificates::BlockCertificates::try_from(response.into_inner())?;
         Ok(QueryResponse {
             block_hash,
             response,