+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased changes
+- Add support for `DryRun`. Requires a node version at least 6.2.
 - Add a `commission_rates` field to `CurrentPaydayBakerPoolStatus` which yields the commission rates
   of the baker for the reward period. Requires a node version at least 6.1.
 - Add support for `GetWinningBakersEpoch`. Requires a node version at least 6.1.